"check_style_score -> 5 of 5 passed. style score: 0/1 "

+++ CSPP2-assignment/M2/Assignment-3/Solution.java
@@ -10,15 +10,15 @@ public final class Solution {
 	}
 	public static void main(final String[] args) {
 		Scanner s=new Scanner(System.in);      
-        int base = s.nextInt();
-        int exponent = s.nextInt();
-        int result=power(base, exponent);
+        long base = s.nextInt();
+        long exponent = s.nextInt();
+        long result=power(base, exponent);
         System.out.println(result);
 	}
 	/*
 	Need to write the power function and print the output.
 	*/
-	public static int power(int base, int exponent) {
+	public static long power(long base, long exponent) {
 		if (exponent==0)
 			return 1;
 		else
